# api-impl/src/main/resources/api-spec.yaml
openapi: '3.0.3'
info:
  version: '0.1.0'
  title: Fruits API
  description: An API to demonstrate Swagger and Spring Boot
security:
  #  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /pointScales:
    post:
      description: Create a new point scale
      operationId: createPointScale
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/PointScale'
      responses:
        '201':
          description: success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PointScale'

    get:
      description: get the list of all point scales
      operationId: getPointScales
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PointScale'
  /pointScales/{pointScale}/leaderBoard:
    get:
      description: Get the leader board for the given point scale
      operationId: getLeaderBoard
      parameters:
        - in: path
          name: pointScale
          schema:
            type: string
          required: true
          description: Name of the point scale in the application
        - in: query
          name: limit
          schema:
            type: integer
          description: The number of users to return in the leader board
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/LeaderBoard'
...
  /events:
    post:
      description: Trigger a new event
      operationId: triggerEvent
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '201':
          description: success
        '400':
          description: Bad request, must have a valid eventType
components:
  securitySchemes:
    #    bearerAuth:            # arbitrary name for the security scheme
    #      type: http
    #      scheme: bearer       # warning: it HAS TO be a lowercase b for the swagger ui version used in springfox
    #      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  schemas:
...
    PointScale:
      type: object
      properties:
        name:
          type: string
    UserScore:
      type: object
      properties:
        appUserId:
          type: string
        score:
          type: int
    LeaderBoard:
      type: object
      properties:
        pointScale:
          type: string
        leaders:
          type: array
          items:
            $ref: '#/components/schemas/UserScore'
...
    Event:
      type: object
      properties:
        eventType:
          type: string
        appUserId:
          type: string